requirements.txt
#requirements.txt
# Telegram Bot
aiogram==3.4.1

# Settings & Config
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1

# Scheduler
APScheduler==3.10.4

# HTTP & Parsing
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==5.1.0

# Selenium for WB cookies
undetected-chromedriver==3.5.5
selenium==4.16.0

# Required for Python 3.12+
setuptools==80.9.0

# Utilities
urllib3==2.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6

playwright==1.40
